Ensure player coordiantes are saved on shutdown
First attempt at ensuring that if a player never makes it to a server
and it is shut down that their coordinates are saved correctly.
Additionally fixed bug in queue were you might be added multiple times
to a queue.
Fixed bug where if you quit and came back the lobby would be your
reconnect server, now it correctly tries to place you in queue if it is
full.